requirements.txt
alembic==0.9.1
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Cerberus==0.8.1
click==6.7
colorama==0.3.7
Eve==0.5.3
Eve-SQLAlchemy==0.4.1
Eve-Swagger==0.0.7
Events==0.2.2
Flask==0.10.1
Flask-Migrate==2.0.3
Flask-PyMongo==0.3.1
Flask-Script==2.0.5
Flask-Script-Extras==0.1.4
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.5
kombu==4.0.2
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.0
pymongo==2.9.4
pyparsing==2.2.0
python-editor==1.0.3
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.6
sqlalchemy-mixins==0.2.2
sqlparse==0.2.3
typing==3.6.1
vine==1.1.3
Werkzeug==0.10.4
